requirements.txt
django
psycopg2-binary
gunicorn
python-dotenv
requests
django-encrypted-model-fields